Move configuration into separate module
